Tea-9656: Splitt opp enhetstester og integrasjonstester
