+++ src/test/integrasjonstester/kotlin/DevLauncherPostgres.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ks.sak
 
 import no.nav.familie.ks.sak.config.ApplicationConfig
-import no.nav.familie.ks.sak.database.DbContainerInitializer
+import no.nav.familie.ks.sak.config.DbContainerInitializer
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.builder.SpringApplicationBuilder
 
@@ -11,16 +11,7 @@ class DevLauncherPostgres
 fun main(args: Array<String>) {
     System.setProperty("spring.profiles.active", "postgres")
     val springBuilder = SpringApplicationBuilder(ApplicationConfig::class.java).profiles(
-        "dev",
-        "postgres",
-        "mock-brev-klient",
-        "mock-økonomi",
-        "mock-infotrygd-feed",
-        "mock-infotrygd-barnetrygd",
-        "mock-pdl",
-        "mock-ident-client",
-        "mock-tilbakekreving-klient",
-        "task-scheduling"
+        "local, mock-pdl, mock-oauth, mock-oppgave, mock-integrasjoner, mock-økonomi"
     )
 
     if (args.contains("--dbcontainer")) {
